% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_data.R
\docType{data}
\name{birth_data}
\alias{birth_data}
\title{Descriptive measurements for eight strains of mice}
\format{
A tibble with 32 rows and 4 variables:
\describe{
  \item{ID}{a factor denoting mouse sex and identification number}
  \item{Treatment}{a factor denoting the treatment received (Placebo, Treatment)}
  \item{sex}{a factor denoting mouse sex (female, male)}
  \item{Num}{an integer corresponding to the identification number}
}
}
\source{
{Stat 108 Professor James Xenakis https://canvas.harvard.edu/courses/140628/files/folder/Projects?}
}
\usage{
birth_data
}
\description{
Includes measurements for treatment, sex, and identification.
This is a subset of \code{\link{data_raw}}.
}
\keyword{datasets}
