% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyweight_summary.R
\name{bodyweight_summary}
\alias{bodyweight_summary}
\title{Bodyweight_Summary}
\usage{
bodyweight_summary(merged_data)
}
\arguments{
\item{merged_data}{A dataframe that contains sex, treatment, bodyweight, and id info}
}
\value{
A table containing summary information on Bodyweights at each checkpoint
}
\description{
This function highlights summaries of bodyweight data.
}
