% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/female_data.R
\docType{data}
\name{female_data}
\alias{female_data}
\title{Body weight measurements for female mice on different diets}
\format{
A tibble with 24 rows and 2 variables:
\describe{
  \item{Diet}{an indicator denoting the diet the mouse was on}
  \item{Bodyweight}{a float indicating the weight of the mouse}
}
}
\source{
{Kaggle, by user Evagoras Georgiou. Link: https://www.kaggle.com/datasets/evagorasgeorgiou/miceweights/data}
}
\usage{
female_data
}
\description{
Includes measurements for female mice bodyweight and indications of treatment.
}
\keyword{datasets}
