% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gender_mismatches.R
\name{find_gender_mismatches}
\alias{find_gender_mismatches}
\title{Find Gender Mismatch}
\usage{
find_gender_mismatches(data = merged_data)
}
\arguments{
\item{data}{A dataframe that contains sex, treatment, bodyweight, and id info, default is merged_data.}
}
\value{
A tibble containing rows where `Sex` does not match `Sex_inferred`.
}
\description{
This function identifies gender mismatches between sex and sex_inferred.
}
