% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_table.R
\name{generate_summary_table}
\alias{generate_summary_table}
\title{Generate a Summary Table of Average Body Weights}
\usage{
generate_summary_table(data)
}
\arguments{
\item{data}{A cleaned and ordered data frame from `clean_mouse_data`.}
}
\value{
A summary table (tibble) showing average body weights by treatment and sex.
}
\description{
Generate a Summary Table of Average Body Weights
}
