% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_body_weight.R
\name{plot_body_weight}
\alias{plot_body_weight}
\title{Plot Mouse Body Weight Over Time by Treatment Group}
\usage{
plot_body_weight(data)
}
\arguments{
\item{data}{A cleaned and ordered data frame from `clean_mouse_data`.}
}
\value{
A ggplot object showing body weight trends over time.
}
\description{
Plot Mouse Body Weight Over Time by Treatment Group
}
