% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mouse_weight.R
\name{plot_mouse_weight}
\alias{plot_mouse_weight}
\title{Plot Mouse Weight}
\usage{
plot_mouse_weight(data = merged_data)
}
\arguments{
\item{data}{A dataframe that contains sex, treatment, bodyweight, and id info, default is merged_data}
}
\value{
A plot highlight the change in mouse weight over time
}
\description{
This function plots the bodyweight of mice over time
}
