% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary.R
\name{stat_summary}
\alias{stat_summary}
\title{Stat_summary}
\usage{
stat_summary(data = merged_data)
}
\arguments{
\item{data}{A dataframe that contains sex, treatment, bodyweight, and id info, default is merged_data}
}
\value{
A list containing the statistical summary table, t-test p-value (comparing Bodyweight 1 and 3), and ANOVA p-value (for 3 bodyweights).
}
\description{
This function outputs statistical summaries regarding bodyweight data for the mice.
}
