% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_test.R
\name{visualize_outliers}
\alias{visualize_outliers}
\title{Provides a visualization the numeric data}
\usage{
visualize_outliers(df2, ..., z_threshold = 3)
}
\arguments{
\item{df2}{A NFL Dataset}

\item{...}{Dataframe columns}

\item{z_threshold}{Cutoff for outlier detection}
}
\value{
returns a plot that shows the distribution of each of the columns, and computes outliers using z-score and plots in red
}
\description{
Provides a visualization the numeric data
}
